Implements minimalistic keyboard input
The MCU is now able to recieve data from the X68k keyboard.
The PC88 output pin is used to blink a led whether the "A" key was
pressed, and pin 4 is used to light up a LED when the "A key released"
message is recieved.

The X68k Keyboard ready is not used.

PC88 output and keycode lookup are not implemented yet.