Implements keycode translation logic
A "key matrix", similiar in how the PC-8801 groups its keys by rows, has
been added. When a valid keycode is recieved from the X68k keyboard, the
corresponding bit (if any) is set or cleared in the matching row.

Then, the value of the key matrix for the row is packed with the row
index, then sent to the PC-8801. This logic is sufficient for handling
the "regular" keycodes, however, several keys on the PC-8801 Type-A
keyboards send multiple keycodes (like F6, which sends the code for F6
followed by the Shift and F1 codes). These "extended codes" are not
supported yet.

The X68k-to-PC88 translation matrix can be found in src/keymap.c.